requirements.txt
# Tested on Python 3.12.1

# Core Deep Learning Frameworks
torch==2.8.0 # No CUDA is needed or used on macOS.
# torch==2.3.0+cu121 # Optional: NVIDIA GPU support
torchvision==0.23.0
torchaudio==2.8.0

# Spiking Neural Networks
snntorch==0.8.0 
brian2==2.9.0
spikingjelly==0.0.0.0.14

# Gradient Boosting
xgboost==3.0.5

# Data Preprocessing and ML 
numpy==1.26.4       
pandas==2.2.2       
scikit-learn==1.6.1
imbalanced-learn==0.13.0  # For handling imbalanced datasets

# Visualization
matplotlib==3.9.0   
seaborn==0.13.2     

# Explainable AI (XAI)
shap==0.45.1        
lime==0.2.0.1       
captum==0.8.0     # PyTorch-native library for interpretability and XAI

# Green AI/Efficiency Metrics
codecarbon==3.0.4 # Energy Consumption Estimation
ptflops==0.6.8 # FLOPs estimation for PyTorch models

# Configuration management
pyyaml==6.0.1    
omegaconf==2.3.0 

# Development Utilities 
ipykernel==6.29.4   
jupyterlab==4.2.1   



